+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.pythonPath": "/Users/nils/.local/share/virtualenvs/menu-service-betOYhT6/bin/python",
+    "python.jediEnabled": false
+}
