+++ backend/mlm-backend/mlm_backend/__main__.py
@@ -1,0 +1,10 @@
+from os import path
+
+from api import SUGGESTION_LOGS, run
+
+
+if not path.isfile(SUGGESTION_LOGS):
+    with open(SUGGESTION_LOGS, 'a') as target:
+        target.write(f'Logfile: `{SUGGESTION_LOGS}`.\n')
+
+run()
