+++ backend/mlm-backend/pyproject.toml
@@ -1,0 +1,11 @@
+[tool.poetry]
+name = "mlm-backend"
+version = "0.1.0"
+description = ""
+authors = ["BracketJohn"]
+
+[tool.poetry.dependencies]
+python = "*"
+aiohttp = "3.4.4"
+
+[tool.poetry.dev-dependencies]
