+++ frontend/nuxt.config.js
