--- pages/index.vue
@@ -1,630 +1,0 @@
-<template>
-
-  <section class="hero is-fullheight">
-      <div class="hero-body">
-      <div class="container">
-
-        <div class="columns">
-          <div class="column has-text-right has-text-centered-mobile">
-            <p class="is-size-3">
-              Is
-            </p>
-          </div>
-          <div class="column has-text-centered">
-            <b-field>
-              <b-input size="is-large" placeholder="Company Name" v-model="companyName" rounded></b-input>
-            </b-field>
-
-            <div v-if="companyName">
-              <div v-if="isMlm">
-                <h1 class="is-size-2 has-text-danger">Yes!</h1>
-                <h2>Please check out <a href="https://en.wikipedia.org/wiki/Multi-level_marketing">Wikipedia</a>, <a href="http://www.pinktruth.com/">Pink truth</a>, <a href="https://www.youtube.com/watch?v=s6MwGeOm8iI&feature=youtu.be">John Oliver's video</a> or <a href="https://www.reddit.com/r/MLMRecovery">this awesome subreddit</a> for information and help!</h2>
-              </div>
-              <div v-else>
-                <h1 class="is-size-2 has-text-success">No!</h1>
-                <h2 v-if="getSimilarCompanyNames">Perhaps you meant: {{ getSimilarCompanyNames }}?</h2>
-              </div>
-            </div>
-            <div v-else>
-              <h1 class="is-size-4 is-size-6-mobile">Please type in a company name!</h1>
-            </div>
-
-          </div>
-          <div class="column has-text-left has-text-centered has-text-centered-mobile">
-            <p class="is-size-3">
-            an MLM?
-            </p>
-          </div>
-        </div>
-
-      </div>
-      </div>
-
-  </section>
-
-
-</template>
-
-<script>
-const levenshtein = require('fast-levenshtein');
-
-export default {
-  computed: {
-    isMlm() {
-      return this.knownMlms.indexOf(this.companyName.toLowerCase()) > -1;
-    },
-    getSimilarCompanyNames() {
-      let distances = [];
-      this.knownMlms.forEach(element => {
-        let distance = levenshtein.get(element, this.companyName);
-        distances.push({
-          name: element,
-          distance: distance
-        });
-      });
-      return distances.sort(function compare(compOne,compTwo) {
-        if (compOne.distance < compTwo.distance)
-          return -1;
-        if (compOne.distance > compTwo.distance)
-          return 1;
-        return 0;
-      }).slice(0, 3).filter(comp => comp.distance < 5).map(comp => comp.name).join(', ');
-    }
-  },
-  data: function() {
-    return {
-      companyName: '',
-      knownMlms: [
-        'premier jewelry',
-        'premier designs',
-        'nutrilite',
-        'cutco',
-        'vector',
-        '31',
-        'thirty one',
-        '5linx',
-        'abby & anna',
-        'acan pacific',
-        'acn',
-        'actilabs',
-        'adornable.u',
-        'advocare',
-        'aerogrow',
-        'agnes & dora',
-        'aim global',
-        'akasuka ',
-        'alcone',
-        "alice's table",
-        "all'asta",
-        'allysian sciences',
-        'aloe vera of america ',
-        'aloette',
-        'alphay int',
-        'alureve',
-        'amare global',
-        'ambit',
-        'amelia james',
-        'ameo',
-        'american income life',
-        'amsoil',
-        'amway',
-        'ann summers',
-        'anorak',
-        'anran',
-        'apollo ',
-        'aquasource uk',
-        'arbonne',
-        'ariix',
-        'arsoa honsha ',
-        'asea global',
-        'asirvia',
-        'aspire/digital altitude',
-        "athena's",
-        'atomy',
-        'ava anderson',
-        'ava rose',
-        'avon',
-        'b:hip global',
-        'bachar nutrition',
-        'bamboo pink',
-        'barefoot books',
-        'bath.ologie',
-        'beach body',
-        'beachbody',
-        'bearcereju ',
-        'beauticontrol',
-        'beauty counter',
-        'beauty society',
-        'because cosmetics',
-        'become international ',
-        'bedroom kandi',
-        'beever',
-        'belcorp ',
-        'bellame',
-        'bemer',
-        'bioperformance',
-        'bod',
-        'body by vi',
-        'body shop at home',
-        'boisset collection',
-        'boston finney ',
-        'bounce life/network',
-        'bud star ',
-        'burnlounge ',
-        'buskins',
-        'butterfly beauty',
-        'cabi',
-        'cambridge weight plan/diet',
-        'can',
-        'captain tortue',
-        'carico int',
-        'celebrating home',
-        'cellements',
-        'chalk couture',
-        'chalky & co',
-        'chandeal ',
-        'charle ',
-        "charlie's project",
-        "chef's toolbox ",
-        'cherish natural products',
-        'chloe & isabel',
-        'clever container',
-        'close to my heart',
-        'cloud 9 parties',
-        'cocoa exchange',
-        'college works',
-        'color by amber',
-        'color happy',
-        'colorstreet',
-        'colour me beautiful',
-        'compelling creations',
-        'conklin',
-        'cookie lee ',
-        'cosway (malaysia)',
-        'country scents',
-        'country suds',
-        'create your life',
-        'creative memories',
-        'credit repair usa',
-        'crunchi',
-        'cutco/vector',
-        'cvsl',
-        'daisy blue naturals',
-        'damsel in defense',
-        'darceys',
-        'dazzle and daze',
-        'diana ',
-        'dione cosmetics',
-        'direct cellars/dc nation',
-        'discovery toys',
-        'divvee/nui',
-        'dot dot smile',
-        'doterra',
-        'du northing designs',
-        'dubli network',
-        'dudley beauty',
-        'dxn',
-        'dynamic essentials',
-        'ecowarehouse',
-        'elepreneuer',
-        'elk river soaps',
-        'ella tina',
-        'elli kai',
-        'elvacity',
-        'emgoldex/global intergold',
-        'enagic/kangen water',
-        'endless xpressions',
-        'enersource int',
-        'enjo ',
-        'envy jewelry',
-        'epicure ',
-        'equinox international ',
-        'essential bodywear',
-        'european grouping of marketing professionals/cedipac sa ',
-        'european home retail ',
-        'evanescence network',
-        'ever skincare',
-        'evolution travel',
-        'evolvhealth',
-        'faberlic ',
-        'family first life',
-        'fantasia',
-        'fantasia ',
-        'fes connect',
-        'fibi & clo',
-        'fifth ave collection',
-        'first fitness nutrition',
-        'fitteam global',
-        'flamingo paperie',
-        'fleuresse',
-        'fm world',
-        'for tails only',
-        'forever living',
-        'forex entourage',
-        'fortune hi',
-        'four oceans',
-        'fragant jewels',
-        'freelife',
-        'fuel freedom int',
-        'fund america ',
-        'gano excel',
-        'gelmoment',
-        'gemstra',
-        'genesis pure',
-        'global affiliate zone',
-        'global legacy initiative',
-        'godesana',
-        'gold canyon',
-        'golden days',
-        'grace & heart',
-        'greeting cake company',
-        'h2o at home',
-        'hanky panky parties ',
-        "hayward's gourmet popcorn",
-        'he(l)o',
-        'healthstyle emporium',
-        'healthy peach',
-        'heavenly chia',
-        'heka corp',
-        'helo wristbands',
-        'hempworx',
-        'herbalife',
-        'heritage makers',
-        'hibody',
-        'hinode',
-        'holiday magic ',
-        'home interiors',
-        'honey',
-        'honey & lace',
-        'hualin biotech',
-        'icoinpro',
-        'id life',
-        'igniting passion ',
-        'imarketslive',
-        'in a pikle',
-        'income advantage',
-        'india hicks',
-        'infinitus',
-        'initials, inc',
-        'inkd up nails',
-        'innov8tive nutrition',
-        'inteletravel',
-        'intimo ',
-        'isagenix',
-        'itworks!',
-        'j. elizabeth',
-        'j. hilburn',
-        'j.r watkins',
-        'jafra',
-        'jamberry',
-        'jamby',
-        'jamie at home ',
-        'janice collection',
-        'java momma',
-        'javita',
-        'jbloom',
-        'jequiti',
-        'jerky direct',
-        'jeunesse',
-        'jewel kade ',
-        'jewelscent',
-        'jk apparel ',
-        'jordan essentials',
-        'joymain',
-        'joyome ',
-        'juiceplus',
-        'jump natural',
-        'kaesar & blair',
-        'kalaia',
-        'kalo & co',
-        'kannaway',
-        'karat bars',
-        'kaszazz',
-        'keep collective',
-        'keep me safe',
-        'keto ',
-        'keto coffee',
-        'ketones',
-        'kirby',
-        'kleeneze',
-        'kobold',
-        'kyani',
-        "l'bri",
-        'labella baskets',
-        'lady godiva beauty',
-        'lavylites',
-        'le',
-        'legal shield',
-        'legging army',
-        'legging girl',
-        'lemongrass spa',
-        'lereve ',
-        'lia sophia',
-        'life abundance',
-        'life leadership',
-        'life tree world',
-        "life's abundance",
-        'lifebrook',
-        'lifevantage',
-        'lilla rose',
-        'limelife',
-        'limu',
-        'linen world',
-        'lipsense',
-        'liv international',
-        'live sore',
-        'longrich',
-        'lorraine lee linen',
-        'love winx',
-        'lr beauty & health',
-        'lularoe',
-        'lulu ave',
-        'luminess',
-        'lyoness',
-        'm. global ',
-        'm. network',
-        'maelle beauty',
-        'magnabilities',
-        'magnolia & vine',
-        'makeup eraser',
-        'man cave',
-        'mannatech',
-        'mark.',
-        'market america',
-        'marly ray',
-        'marvelous mouse travels',
-        'mary & martha',
-        'marykay',
-        'mary kay',
-        'maskara',
-        'matilda jane',
-        'max & madeleine',
-        'maxwell clothing',
-        'mca',
-        'medifast',
-        'melaleuca',
-        'metabolife ',
-        'mia bath and body ',
-        'mialisia',
-        'miche eu',
-        'miki',
-        'moa nutrition',
-        'modere/neway',
-        'mojilife',
-        'monat',
-        'monavie ',
-        'morinda bioactives',
-        'motives cosmetics',
-        'multpure',
-        'my club 8',
-        'my daily choice',
-        'my lala leggings',
-        'myecon',
-        'national safety associates',
-        'national wealth center',
-        'natura ',
-        'nature direct ',
-        "nature's sunshine products",
-        "neal's yard remedies organic",
-        'neolife',
-        'nerium',
-        'nevetica',
-        'new era',
-        'new u life',
-        'neway/modere',
-        'nikken',
-        'noevir',
-        'nomades',
-        'noonday collection',
-        'norwex',
-        'nouveau riche  ',
-        'nspire network',
-        'nucerity',
-        'nuskin',
-        'nxivm',
-        'omnilife',
-        'one hope wine',
-        'optavia',
-        'opulenza',
-        'organo gold',
-        'oriflame',
-        'origami owl',
-        'our hearts desire',
-        'paid 2 save',
-        'pampered chef',
-        'paparazzi',
-        'paperly',
-        'park lane jewelry',
-        'party girl',
-        'party lite',
-        'party time mixes',
-        'partylite',
-        'passion parties',
-        'pawtree',
-        'paycation',
-        'peach',
-        'pearl chic',
-        'peekaboo beans',
-        'perfect',
-        'perfectly polished',
-        'perfectly posh',
-        'personally poetic',
-        'php',
-        'pink zebra',
-        'piphany',
-        'pixielane',
-        'plexus',
-        'plunder',
-        'pm international',
-        'pola ',
-        'poofy organics',
-        'premier designs',
-        'premier financial',
-        'primemybody',
-        'primerica',
-        'princess house',
-        'prodoula',
-        'proyoung',
-        'pruvit',
-        'pulse cosmetics',
-        'pure haven',
-        'pure romance',
-        'purehaven',
-        'purely',
-        'purium',
-        'qivana',
-        'qnet',
-        'quanjian natural',
-        'rain international',
-        'rainbow vacuum',
-        'real time pain relief',
-        'red aspen',
-        'red safety',
-        'reliv',
-        'revitalu',
-        'riway',
-        'rodan+fields',
-        'roland',
-        'rolmex',
-        'royal tongan limu ',
-        'royaltie gens',
-        'ruby ribbon',
-        'saba',
-        'sabika jewelry',
-        'safegirl security',
-        'salad master',
-        'sarso ',
-        'scentsy',
-        "schneider's gourmet world",
-        'scout & cellar',
-        'seacret',
-        'sendoutcards',
-        'senegence',
-        'shakeology',
-        'shaklee',
-        'shopping sherlock',
-        'shrimp & grits',
-        'signature homestyles',
-        'silpada',
-        'silver icing',
-        'simple man',
-        'simply success elite',
-        'simplyfun games',
-        'skinny body at home',
-        'solavei',
-        'solvei ',
-        'sophie paris ',
-        'south hill designs',
-        'southern living at home',
-        'southwestern advantage',
-        'sseko',
-        'stamped 4 u',
-        'stampin up',
-        'steam energy',
-        'steeped tea',
-        'stella & dot',
-        'stream energy',
-        'style dots',
-        'success university',
-        'sunrider',
-        'sunset gourmet',
-        'sunshine empire ',
-        'surge 365',
-        'sweet legs',
-        'sweet minerals',
-        'symmetry financial group',
-        'syntek global',
-        'tastefully simple',
-        'tavala',
-        'tealightful',
-        'team national',
-        'tedivina',
-        'telecom plus',
-        'telexfree ',
-        'the advert platfrom',
-        'the body shop at home',
-        'the landmark forum',
-        'the longabeger company',
-        'the super affiliate network',
-        'thermomix',
-        'thrive',
-        'thrive life',
-        'tiber river naturals',
-        'tocara ',
-        'touchstone crystal',
-        'touchstone essentials',
-        'trades of hope',
-        'tranont',
-        'transformational beauty',
-        'travel evolution',
-        'traveling vineyard',
-        'traverus global',
-        'trivita',
-        'tropic skin care',
-        'truvision health',
-        'ts',
-        'tupperware',
-        'united sciences of america',
-        'united warehouse',
-        'univera',
-        'usana',
-        'usborne',
-        'utility warehouse',
-        'valentus',
-        'vantel',
-        'vasayo',
-        'vector/cutco',
-        'vemma',
-        'viai la teeth whitening',
-        'viaonehope',
-        'vibella',
-        'vida divina',
-        'vie at home',
-        'visalus ',
-        'vitality extracts',
-        'vivamk',
-        'volo',
-        'vorwerk',
-        'votre belle maison',
-        'voxxlife',
-        'wakaya perfection',
-        'wakeupnow',
-        'watkins inc',
-        'wealthperx',
-        'wikaniko',
-        'wildtree',
-        'willing beauty',
-        'winasun',
-        'wine shop at home',
-        'wink naturals',
-        'world financial group',
-        'world leadership group',
-        'world ventures',
-        'wotabu',
-        'xango',
-        'xerveo',
-        'xoom energy',
-        'xstream travel',
-        'xyngular',
-        'yanbal int',
-        'yandi',
-        'yelloow',
-        'yevo',
-        'yofoto',
-        'yoli',
-        'yoonla',
-        'young living',
-        'youngevity',
-        'younique',
-        'ytb international',
-        'zija',
-        'zilis',
-        'zinzino',
-        'zrii',
-        'zurvita',
-        'zyia',
-        'zyn']
-    }
-  },
-}
-</script>
